# docs/api/openapi-full.yaml
openapi: 3.1.0
info:
  title: DataSafeS3 REST Admin API
  version: 1.1.0
  summary: Complete JSON REST API served at /api/v1 (excluding S3 XML)
  description: |
    **Full hand-written OpenAPI 3.1** for all Admin JSON routes registered in internal/api/server.go.

    ## Authentication
    | Scheme | Use |
    |--------|-----|
    | **BearerAPIToken** | Integration tokens ds_* (recommended for automation) |
    | **BearerJWT** | JWT from POST /admin/login or OIDC callback (console sessions) |

    Public endpoints: GET /health, POST /admin/login, OIDC bootstrap, GET /public/share/*.

    ## Not included
    - GET /metrics (Prometheus text)
    - S3 XML API on paths outside /api/v1 (SigV4, port 9000)

    ## Errors
    Standard envelope: JSON object with error field (HTTP 4xx/5xx).

    Interactive docs for **Community subset**: GET /api/v1/docs (Swagger UI).

    Canonical file: [openapi-full.yaml](../../api/openapi-full.yaml)
  contact:
    name: DataSafeS3 / Ilya Trachuk
    url: https://github.com/DirektorBani/DataSafeS3
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Via Caddy (console origin)
  - url: http://localhost:9000/api/v1
    description: storage-server direct
security:
  - BearerAPIToken: []
  - BearerJWT: []
components:
  securitySchemes:
    BearerAPIToken:
      type: http
      scheme: bearer
      bearerFormat: ds_
      description: |
        Integration API token with prefix ds_. Create in console: Access → API tokens.
        Inherits creator's role and bucket access.

    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT from POST /admin/login, MFA completion, or OIDC callback.
  parameters:
    BucketName:
      name: bucket
      in: path
      required: true
      schema: { type: string }
    ObjectKey:
      name: key
      in: path
      required: true
      schema: { type: string }
      description: Object key (may contain slashes)
...
    Prefix:
      name: prefix
      in: query
      schema: { type: string }
    Delimiter:
      name: delimiter
      in: query
      schema: { type: string, example: "/" }
    StartAfter:
      name: start_after
      in: query
      schema: { type: string }
    MaxKeys:
      name: max_keys
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        mfa_required: { type: boolean }
        mfa_token: { type: string }
        mfa_setup_required: { type: boolean }
        scheduled_delete_at: { type: string, format: date-time }
        trashed: { type: boolean }
        trash_id: { type: string }
...
    Object:
      type: object
      properties:
        bucket: { type: string }
        key: { type: string }
        size: { type: integer, format: int64 }
        etag: { type: string }
        content_type: { type: string }
        version_id: { type: string }
        last_modified: { type: string, format: date-time }
        legal_hold: { type: boolean }
        retention_until: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: { type: string } }
        tags: { type: object, additionalProperties: { type: string } }
    ObjectListResponse:
      type: object
      properties:
        objects:
          type: array
          items: { $ref: '#/components/schemas/Object' }
        folders:
          type: array
          items: { type: string }
        truncated: { type: boolean }
        next_marker: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /buckets/{bucket}/objects:
    get:
      operationId: listObjects
      summary: List objects
      tags: [Buckets]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/BucketName'
          - $ref: '#/components/parameters/Prefix'
          - $ref: '#/components/parameters/Delimiter'
          - $ref: '#/components/parameters/StartAfter'
          - $ref: '#/components/parameters/MaxKeys'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /buckets/{bucket}/objects/{key}:
    get:
      operationId: downloadObject
      summary: Download object
      tags: [Buckets]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/BucketName'
          - $ref: '#/components/parameters/ObjectKey'
      responses:
        '200':
          description: Object bytes
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    put:
      operationId: uploadObject
      summary: Upload object
      tags: [Buckets]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/BucketName'
          - $ref: '#/components/parameters/ObjectKey'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { $ref: '#/components/schemas/Object' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: deleteObject
      summary: Delete or schedule delete object
      tags: [Buckets]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/BucketName'
          - $ref: '#/components/parameters/ObjectKey'
          - name: schedule
            in: query
            schema: { type: string, enum: [1d, 1w, 1m] }
          - name: versionId
            in: query
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '200': { description: Scheduled delete or soft-delete to trash }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /gateway/replication:
    get:
      operationId: listReplicationRules
      summary: List replication rules
      tags: [Gateway]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Rules }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: createReplicationRule
      summary: Create replication rule
      tags: [Gateway]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /gateway/replication/{id}:
    delete:
      operationId: deleteReplicationRule
      summary: Delete replication rule
      tags: [Gateway]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /site-replication/peers:
    get:
      operationId: listSiteReplicationPeers
      summary: List site replication peers
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Peers }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: createSiteReplicationPeer
      summary: Register DataSafeS3 peer site
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /site-replication/peers/{id}:
    delete:
      operationId: deleteSiteReplicationPeer
      summary: Remove site replication peer
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /site-replication/rules:
    get:
      operationId: listSiteReplicationRules
      summary: List site replication rules
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Rules }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: createSiteReplicationRule
      summary: Create site replication rule
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /site-replication/rules/{id}:
    delete:
      operationId: deleteSiteReplicationRule
      summary: Delete site replication rule
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /site-replication/status:
    get:
      operationId: siteReplicationStatus
      summary: Site replication lag and queue
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Status }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...